Add "release" `uswds` stylesheets;
This patch adds the "released" uswds compiled stylesheet. This is a
replacement of the `scss` files. Rather than include the `scss` source
files in the ruby gem, the compiled CSS is the only thing including.
This is still up for discussion, as Rails and Jekyll both have their own
asset pipelines and can handle having the source files to compile. Not
having the source files though does help the standard be more
authoritative and prevents developer confusion over what `scss` files to
include and which to ignore.