Add "release" `uswds` JS
This patch adds the "released" uswds compiled JavaScript. This is a
refactor of the old `components.js` file which included everything in
one file. This file has been bundled using `browserify`. The original
CommonJS files are not included. This is still up for discussion. Not
having the source files helps prevent developer confusion surrounding
what components to choose from. This JavaScript bundle _is opinionated_
and ships with the version of jQuery required by the `uswds` which is
jQuery 1.11.3 as of this build.